+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `secrets+*-import://` added syntax checks 
+
 ## [4.5.0] - 2023-08-14
 
 ### Added